# public/.well-known/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MegaChad Agent API",
    "version": "1.0.0",
    "description": "Burn-to-create looksmaxxing engine on MegaETH. Burn $MEGACHAD tokens to generate AI-enhanced portraits, mint NFTs, and earn on-chain reputation. All endpoints run on MegaETH (chain ID 4326, RPC: https://mainnet.megaeth.com/rpc).",
    "contact": {
      "name": "MegaChad",
      "url": "https://megachad.xyz"
    }
  },
  "servers": [
    {
      "url": "https://megachad.xyz",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/defi/staking/tx": {
      "get": {
        "operationId": "buildStakingTx",
        "summary": "Build approve+stake / unstake / claim transaction",
        "description": "Returns `executionPlan: [{ to, value, data, functionName, args }]`. For stake, the plan starts with approve if allowance is insufficient (or always, if `address` is omitted).",
        "parameters": [
          { "name": "action", "in": "query", "required": true, "schema": { "type": "string", "enum": ["stake", "unstake", "claim"] } },
          { "name": "venue", "in": "query", "required": true, "schema": { "type": "string", "enum": ["mogger", "jester"] } },
          { "name": "amount", "in": "query", "required": false, "schema": { "type": "string" }, "description": "Required for stake/unstake; human units (e.g. \"1000\")." },
          { "name": "address", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Ordered action plan" } }
      }
    },
...
    "/api/defi/amm/swap-tx": {
      "get": {
        "operationId": "buildAmmSwapTx",
        "summary": "Build a MEGACHAD↔MEGAGOONER swap transaction",
        "description": "Returns a two-step plan: ERC20.transfer(pair, amountIn) → pair.swap(amountAIn, amountBIn, to). MegaChadLP has no on-chain minOut — verify off-chain.",
        "parameters": [
          { "name": "from", "in": "query", "required": true, "schema": { "type": "string", "enum": ["MC", "MG"] } },
          { "name": "amount", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "recipient", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "slippageBps", "in": "query", "required": false, "schema": { "type": "integer", "default": 200 } }
        ],
        "responses": { "200": { "description": "Swap execution plan + minOut" } }
      }
    },
...
    "/api/defi/governance/vote-tx": {
      "get": {
        "operationId": "buildVoteTx",
        "summary": "Build castVote transaction",
        "parameters": [
          { "name": "proposalId", "in": "query", "required": true, "schema": { "type": "integer" } },
          { "name": "support", "in": "query", "required": true, "schema": { "type": "string", "enum": ["for", "against", "abstain"] } }
        ],
        "responses": { "200": { "description": "castVote plan" } }
      }
    },
...
  },
...
}